requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
confusable-homoglyphs==3.2.0
crispy-bootstrap5==2023.10
cryptography==41.0.5
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.7
django-allauth==0.58.2
django-configurations==2.5
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-registration==3.4
djangorestframework==3.14.0
drf-yasg==1.21.7
flake8==6.1.0
idna==3.4
inflection==0.5.1
isort==5.12.0
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.2
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing-extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
